Rework testing of cache_data
Mocking `File` was causing unrelated tests to fail.  Switching to
rspec-mocks to mock the File reads was also nasty (I couldn't get it
working anyway!)

I also tried using FakeFS, but that didn't get far because of
https://github.com/fakefs/fakefs/issues/326

The reason that the mocks were needed in the first place was because of
how rspec-puppet works.  It sets up and destroys a temporary puppet
vardir for every example run.  The cache_data function caches data in a
directory under vardir.  If you called the function twice, it'd never
find the cached data, because the second call was looking in a brand new
temporary vardir.

My solution is to use rspec-mocks to replace rspec-puppet's call to
mktmpdir and return our own temporary directory to use as the vardir.