This PR adds a new fact puppet_config
This new fact aims to expose some of the configurations parameters of
puppet.  it is similar to the puppet_settings fact in stdlib but uses
structured facts and exposes a few more settings.

I first attempted to submit something similar to stdlib[1] but got
redirected here.  happy to update if this is useful or abandon if not

[1]https://github.com/puppetlabs/puppetlabs-stdlib/pull/1063